// src/schemas.ts
export const WorldBankProcNoticeRawSchema = z.object({
    id: z.string(),
    notice_type: z.string().nullable().optional(),
    noticedate: z.string().nullable().optional(),
    notice_lang_name: z.string().nullable().optional(),
    notice_status: z.string().nullable().optional(),
    submission_deadline_date: z.string().nullable().optional(),
    submission_deadline_time: z.string().nullable().optional(),
    project_ctry_name: z.string().nullable().optional(),
    project_id: z.string().nullable().optional(),
    project_name: z.string().nullable().optional(),
    bid_reference_no: z.string().nullable().optional(),
    bid_description: z.string().nullable().optional(),
    procurement_group: z.string().nullable().optional(),
    procurement_method_code: z.string().nullable().optional(),
    procurement_method_name: z.string().nullable().optional(),
    contact_address: z.string().nullable().optional(),
    contact_ctry_name: z.string().nullable().optional(),
    contact_email: z.string().nullable().optional(),
    contact_name: z.string().nullable().optional(),
    contact_organization: z.string().nullable().optional(),
    contact_phone_no: z.string().nullable().optional(),
    contact_web_url: z.string().nullable().optional(),
    submission_date: z.string().nullable().optional(),
    notice_text: z.string().nullable().optional(),
});

export type WorldBankProcNoticeRaw = z.infer<typeof WorldBankProcNoticeRawSchema>;